Update QV compile tests to use a real QV circuit
The previous QV transpile benchmarks were all loading the circuit from
qasm2.0 files. However, a real quantum volume circuit should be defined
with haar random SU(4) matrices and the compiler's job is to synthesize
that to a realization that is executable on the hardware. With the
exception of `staq` all the toolkit's we're running this benchmark on
support representing a quantum circuit like this, and we already had
code to generate a quantum volume circuit for construction benchmarks.
Staq is the exception here as it is unable to represent an arbitrary 2q
unitary operator as it only supports qasm 2.0 as the input format. This
commit makes this test a hard failure as the library lacks support for
this where everything else that has a transpiler supports it.