Qiskit 2.0 compatibility
With the pending release of Qiskit 2.0 at the end of this month there
are a few small api changes that the benchpress code needs to account
for. The first is that the `BackendConfiguration` objects that are used
for interop with tket are no longer in Qiskit because they are IBM
Quantum API payload container objects. So instead these are imported
from the IBM client library package qiskit-ibm-runtime. The second
change is that Qiskit 2.0 no longer supports `.c_if()` which changes how
QASM2 is parsed. By default if a condition is contained in a qasm2
program this gets converted into a Qiskit `IfElseOp`. However the
flexible backends used for compilation targets were not indicating they
supported IfElseOp. This was causing an error for the few qasm files
using conditions because the transpilation would fail because if else
wasn't a supported operation. This fixes this by enabling control flow
ops in the felxible backend when used by Qiskit.