Fix compatibility with pytest 8.4.0
This commit updates the benchpress JSON generation to be compatible
with 8.4.0. 8.4.0 changed the internal API for storing the session start
time which broke how benchpress was querying the data. The session start
time doesn't seem to be queriable (at least for my brief 3 min search)
from a public api of the terminal reporter so this opts to just update
the internal usage based on the installed pytest version.